real_estate_tracker/backend/requirements.txt
# Core dependencies
typer[all]>=0.9.0
sqlalchemy>=2.0.0
pandas>=1.5.0
plotly>=5.0.0
openpyxl>=3.1.0
passlib[bcrypt]>=1.7.4
rich>=13.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Security dependencies
cryptography>=41.0.0
pyjwt>=2.8.0
python-multipart>=0.0.6

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
bandit>=1.7.5
safety>=2.3.5

# Database
alembic>=1.13.0

# Utilities
python-dateutil>=2.8.0
click>=8.1.0 